chore: removed extra order details table
