+++ server/db/models/index.js
@@ -15,8 +15,8 @@ const Cart = require('./cart')
 User.hasOne(Cart)
 Cart.belongsTo(User)
 
-Product.belongsToMany(Cart, {through: 'OrderDetail'})
-Cart.belongsToMany(Product, {through: 'OrderDetail'})
+Product.belongsToMany(Cart, {through: 'orderDetail'})
+Cart.belongsToMany(Product, {through: 'orderDetail'})
 
 /**
  * We'll export all of our models here, so that any time a module needs a model,
